+++ OpenMEEG/libs/OpenMEEGMaths/src/symmatrix.cpp
@@ -159,9 +159,9 @@ namespace OpenMEEG {
         // LU
         int *pivots=new int[nlin()];
         int Info = 0;
-        DSPTRF('U',A.nlin(),A.data(),pivots,Info);
+        DSPTRF('U',(int)A.nlin(),A.data(),pivots,Info);
         // Solve the linear system AX=B
-        DSPTRS('U',A.nlin(),RHS.ncol(),A.data(),pivots,RHS.data(),A.nlin(),Info);
+        DSPTRS('U',(int)A.nlin(),(int)RHS.ncol(),A.data(),pivots,RHS.data(),(int)A.nlin(),Info);
         om_assert(Info == 0);
         return RHS;
     #else
